+++ bee-message/src/output/mod.rs
@@ -28,9 +28,9 @@ use core::{
 };
 
 /// The maximum number of outputs of a transaction.
-pub const OUTPUT_COUNT_MAX: usize = INPUT_COUNT_MAX;
+pub const OUTPUT_COUNT_MAX: u16 = INPUT_COUNT_MAX;
 /// The range of valid numbers of outputs of a transaction .
-pub const OUTPUT_COUNT_RANGE: RangeInclusive<usize> = 1..=OUTPUT_COUNT_MAX; //[1..127]
+pub const OUTPUT_COUNT_RANGE: RangeInclusive<u16> = 1..=OUTPUT_COUNT_MAX; //[1..127]
 /// The range of valid indices of outputs of a transaction .
 pub const OUTPUT_INDEX_RANGE: Range<u16> = 0..OUTPUT_COUNT_MAX as u16; //[0..126]
